+++ src/main/java/com/guerrero/Inventario/repository/VentaRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -32,7 +33,7 @@ public interface VentaRepository extends JpaRepository<Venta, Long> {
     @Query("SELECT COALESCE(SUM(v.total), 0) FROM Venta v " +
             "WHERE v.estado = com.guerrero.Inventario.model.Venta.EstadoVenta.COMPLETADA " +
             "AND v.fecha BETWEEN :inicio AND :fin")
-    Double totalVendidoEntre(@Param("inicio") LocalDateTime inicio,
+    BigDecimal totalVendidoEntre(@Param("inicio") LocalDateTime inicio,
                              @Param("fin") LocalDateTime fin);
 
     @Query("SELECT v FROM Venta v WHERE v.sucursal.id = :sucursalId AND v.usuario.id = :usuarioId AND v.fecha >= :inicio AND v.fecha <= :fin")
